kageco-273 freee API取引登録機能をリファクタリングし、コード構造を改善
- submitFreeeDeals関連ロジックを整理し、冗長なコードを削除
- RegisterDealGatewayおよび新しいDTO/RegisterDealDtoを用いる形に変更
- 統一されたインターフェースとデータ構造で処理を行い、可読性と保守性を向上
- 古い関数やテストを削除し、必要な処理をsubmitDealsに集約
