+++ packages/core/gateway/freee/register-deal-gateway.ts
@@ -1,0 +1,13 @@
+import type { RegisterDealDto } from "../../usecase/freee/register-deal-dto";
+
+/**
+ * freee API取引登録ゲートウェイのインターフェース
+ */
+export interface RegisterDealGateway {
+  /**
+   * freeeへ取引データを送信する
+   * @param dealData 取引データ
+   * @returns 成功した場合はtrue、失敗した場合はfalse
+   */
+  exec(dealData: RegisterDealDto): Promise<{ success: boolean }>;
+}
